# Codegen: on demand distribution to forked processes