@ysbaddaden 2024-03-05
Let's note that we default to 8 processes, which used to be fine when we had 4 cores with hyperthreading (8 threads), but might now be much in phase with the current CPU trends where 16, 20 or 32 threads are becoming the norm, though we might not want to fork 160 processes on an Ampere Altra :grin:

Maybe we could improve the default value to be `System.cpu_count.clamp(8..32)` so we have enough processes but not too much (I'm looking at you Ampere Altra).

Maybe we could accept a `CRYSTAL_COMPILER_THREADS` environment variable too?