@ysbaddaden 2024-03-05
And... now I'm wondering about what happens when we have fewer compilation units than configured threads :thinking:

Maybe I should limit the number of forked processes, just in case :sweat_smile: