@ysbaddaden 2024-03-05
I just pushed one last commit to fix edge cases with too little (must be >= 1) or too many processes (pointless).

The comment was to leave some indication about how to go further.