@straight-shoota 2024-03-05
Yeah, that clamp makes sense for sanity.